Feat(#49): rename support/notification to support/fcm
